+++ src/lexer/utils/javascript/index.ts
@@ -2,3 +2,5 @@ export * from "./common";
 export * from "./string";
 export * from "./regex";
 export * from "./numeric";
+export * from "./types";
+export * from "./scanner";