+++ CHANGELOG.md
@@ -9,7 +9,7 @@
   - Feat: add `ITrimmedLexer/IReadonlyTrimmedLexer`.
   - Feat: `Builder.append` can accept actions with different data bindings in one call.
   - Feat: `Builder.ignore` won't update the token data bindings.
-  - Feat: add `Lexer.javascript.evalStringContent`.
+  - Feat: add `Lexer.javascript.evalStringContent/simpleStringLiteral`.
   - Fix: `Lexer.trimStart` shouldn't collect non-muted errors.
 - Parser
   - **_Breaking Change_**: lexer will be trimmed after every parse. This should improve the performance.